# docs/swagger.yaml
definitions:
...
  disgm.Member:
    properties:
      avatar:
        description: The member's guild avatar hash
        type: string
      avatar_decoration_data:
        allOf:
        - $ref: '#/definitions/models.AvatarDecorationData'
        description: Data for the member's guild avatar decoration
      communication_disabled_until:
        description: When the user's timeout will expire
        type: string
      deaf:
        description: Whether the user is deafened in voice channels
        type: boolean
      flags:
        description: Guild member flags represented as a bit set, defaults to 0
        type: integer
      joined_at:
        description: When the user joined the guild
        type: string
      mute:
        description: Whether the user is muted in voice channels
        type: boolean
      nick:
        description: This user's guild nickname
        type: string
      pending:
        description: Whether the user has not yet passed the guild's Membership Screening
          requirements
        type: boolean
      permissions:
        description: Total permissions of the member in the channel, including overwrites
        type: string
      premium_since:
        description: When the user started boosting the guild
        type: string
      roles:
        description: Array of role object IDs
        items:
          type: string
        type: array
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: The user this guild member represents
    type: object
  disgm.Message:
    properties:
      activity:
        description: Sent with Rich Presence-related chat embeds
      application:
        description: Sent with Rich Presence-related chat embeds
      application_id:
        description: ID of the application if the message is an Interaction or application-owned
          webhook
        type: string
      attachments:
        description: Any attached files
        items: {}
        type: array
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: The author of this message (not guaranteed to be a valid user)
      call:
        description: The call associated with the message
      channel_id:
        description: ID of the channel the message was sent in
        type: string
      components:
        description: Sent if the message contains components like buttons, action
          rows, etc.
        items: {}
        type: array
      content:
        description: Contents of the message
        type: string
      edited_timestamp:
        description: When this message was edited (or null if never)
        type: string
      embeds:
        description: Any embedded content
        items: {}
        type: array
      flags:
        description: Message flags combined as a bitfield
        type: integer
      id:
        description: ID of the message
        type: string
      interaction:
        description: Deprecated in favor of interaction_metadata
      interaction_metadata:
        description: Sent if the message is sent as a result of an interaction
      mention_channels:
        description: Channels specifically mentioned in this message
        items: {}
        type: array
      mention_everyone:
        description: Whether this message mentions everyone
        type: boolean
      mention_roles:
        description: Roles specifically mentioned in this message
        items:
          type: string
        type: array
      mentions:
        description: Users specifically mentioned in the message
        items:
          $ref: '#/definitions/models.User'
        type: array
      message_reference:
        description: Data showing the source of a crosspost, channel follow add, pin,
          or reply message
      message_snapshots:
        description: The message associated with the message_reference
        items: {}
        type: array
      nonce:
        description: Used for validating a message was sent
      pinned:
        description: Whether this message is pinned
        type: boolean
      poll:
        description: A poll!
      position:
        description: Approximate position of the message in a thread
        type: integer
      reactions:
        description: Reactions to the message
        items:
          $ref: '#/definitions/models.Reaction'
        type: array
      referenced_message:
        allOf:
        - $ref: '#/definitions/models.Message'
        description: The message associated with the message_reference
      resolved:
        description: Data for users, members, channels, and roles in the message's
          auto-populated select menus
      role_subscription_data:
        description: Data of the role subscription purchase or renewal
      sticker_items:
        description: Sent if the message contains stickers
        items: {}
        type: array
      stickers:
        description: Deprecated the stickers sent with the message
        items:
          $ref: '#/definitions/models.Sticker'
        type: array
      thread:
        allOf:
        - $ref: '#/definitions/models.Channel'
        description: The thread that was started from this message
      timestamp:
        description: When this message was sent
        type: string
      tts:
        description: Whether this was a TTS message
        type: boolean
      type:
        description: Type of message
        type: integer
      webhook_id:
        description: If the message is generated by a webhook
        type: string
    type: object
...
  models.Member:
    properties:
      avatar:
        description: The member's guild avatar hash
        type: string
      avatar_decoration_data:
        allOf:
        - $ref: '#/definitions/models.AvatarDecorationData'
        description: Data for the member's guild avatar decoration
      communication_disabled_until:
        description: When the user's timeout will expire
        type: string
      deaf:
        description: Whether the user is deafened in voice channels
        type: boolean
      flags:
        description: Guild member flags represented as a bit set, defaults to 0
        type: integer
      joined_at:
        description: When the user joined the guild
        type: string
      mute:
        description: Whether the user is muted in voice channels
        type: boolean
      nick:
        description: This user's guild nickname
        type: string
      pending:
        description: Whether the user has not yet passed the guild's Membership Screening
          requirements
        type: boolean
      permissions:
        description: Total permissions of the member in the channel, including overwrites
        type: string
      premium_since:
        description: When the user started boosting the guild
        type: string
      roles:
        description: Array of role object IDs
        items:
          type: string
        type: array
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: The user this guild member represents
    type: object
  models.Message:
    properties:
      activity:
        description: Sent with Rich Presence-related chat embeds
      application:
        description: Sent with Rich Presence-related chat embeds
      application_id:
        description: ID of the application if the message is an Interaction or application-owned
          webhook
        type: string
      attachments:
        description: Any attached files
        items: {}
        type: array
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: The author of this message (not guaranteed to be a valid user)
      call:
        description: The call associated with the message
      channel_id:
        description: ID of the channel the message was sent in
        type: string
      components:
        description: Sent if the message contains components like buttons, action
          rows, etc.
        items: {}
        type: array
      content:
        description: Contents of the message
        type: string
      edited_timestamp:
        description: When this message was edited (or null if never)
        type: string
      embeds:
        description: Any embedded content
        items: {}
        type: array
      flags:
        description: Message flags combined as a bitfield
        type: integer
      id:
        description: ID of the message
        type: string
      interaction:
        description: Deprecated in favor of interaction_metadata
      interaction_metadata:
        description: Sent if the message is sent as a result of an interaction
      mention_channels:
        description: Channels specifically mentioned in this message
        items: {}
        type: array
      mention_everyone:
        description: Whether this message mentions everyone
        type: boolean
      mention_roles:
        description: Roles specifically mentioned in this message
        items:
          type: string
        type: array
      mentions:
        description: Users specifically mentioned in the message
        items:
          $ref: '#/definitions/models.User'
        type: array
      message_reference:
        description: Data showing the source of a crosspost, channel follow add, pin,
          or reply message
      message_snapshots:
        description: The message associated with the message_reference
        items: {}
        type: array
      nonce:
        description: Used for validating a message was sent
      pinned:
        description: Whether this message is pinned
        type: boolean
      poll:
        description: A poll!
      position:
        description: Approximate position of the message in a thread
        type: integer
      reactions:
        description: Reactions to the message
        items:
          $ref: '#/definitions/models.Reaction'
        type: array
      referenced_message:
        allOf:
        - $ref: '#/definitions/models.Message'
        description: The message associated with the message_reference
      resolved:
        description: Data for users, members, channels, and roles in the message's
          auto-populated select menus
      role_subscription_data:
        description: Data of the role subscription purchase or renewal
      sticker_items:
        description: Sent if the message contains stickers
        items: {}
        type: array
      stickers:
        description: Deprecated the stickers sent with the message
        items:
          $ref: '#/definitions/models.Sticker'
        type: array
      thread:
        allOf:
        - $ref: '#/definitions/models.Channel'
        description: The thread that was started from this message
      timestamp:
        description: When this message was sent
        type: string
      tts:
        description: Whether this was a TTS message
        type: boolean
      type:
        description: Type of message
        type: integer
      webhook_id:
        description: If the message is generated by a webhook
        type: string
    type: object
...
  models.Role:
    properties:
      color:
        description: Color of the role
        type: integer
      hoist:
        description: Whether the role is hoisted in the user list
        type: boolean
      id:
        description: Define fields for Role structure based on your needs
        type: string
      managed:
        description: Whether the role is managed by an application
        type: boolean
      mentionable:
        description: Whether the role is mentionable
        type: boolean
      name:
        description: Name of the role
        type: string
      permissions:
        description: Permissions for the role
        type: string
      position:
        description: Position of the role
        type: integer
    type: object
...
  models.Sticker:
    properties:
      description:
        description: Optional description of the sticker
        type: string
      format_type:
        description: Format type of the sticker
        type: integer
      id:
        description: Define fields for Sticker structure based on your needs
        type: string
      name:
        description: Name of the sticker
        type: string
      pack_id:
        description: ID of the sticker pack
        type: string
    type: object
...
host: localhost:90
info:
  contact: {}
  description: API for managing Discord guilds using DiscordGo and Fiber.
  title: Discord Guild Management API
  version: "1.0"
paths:
...
  /api/guild/channels/{channelid}/messages:
    get:
      description: Retrieve all messages from a specific channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/disgm.Message'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Channel Messages
      tags:
      - Messages
    post:
      description: Send a new message to a specific channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Send Channel Message
      tags:
      - Messages
  /api/guild/channels/{channelid}/messages/{messageid}:
    delete:
      description: Delete a specific message in a channel by ID.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete Channel Message
      tags:
      - Messages
    get:
      description: Retrieve a specific message by ID from a channel.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Channel Message
      tags:
      - Messages
    patch:
      description: Edit a specific message in a channel by ID.
      parameters:
      - description: Channel ID
        in: path
        name: channelid
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Edit Channel Message
      tags:
      - Messages
...
  /api/guild/members:
    get:
      description: Retrieve all members of the guild.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/disgm.Member'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Members
      tags:
      - Members
  /api/guild/members/{memberid}:
    get:
      description: Retrieve a specific member from the guild by ID.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Member
      tags:
      - Members
    patch:
      description: Update a specific member in the guild.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update Guild Member
      tags:
      - Members
  /api/guild/members/{memberid}/roles:
    get:
      description: Retrieve all roles assigned to a specific member in the guild.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Member Roles
      tags:
      - Roles
  /api/guild/members/{memberid}/roles/{roleid}:
    delete:
      description: Remove a role from a specific member in the guild.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Remove Member Role
      tags:
      - Roles
    put:
      description: Add a role to a specific member in the guild.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Add Member Role
      tags:
      - Roles
...
  /guilds/{guildid}/members/{memberid}:
    delete:
      description: Remove a member from the specified guild.
      parameters:
      - description: Member ID
        in: path
        name: memberid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Kick Member
      tags:
      - Members
...
swagger: "2.0"